refactor: move rgetattr function to helpers.py
